Reorganize tests by file.
