+++ lib/__tests__/plugin.test.js
