+++ lib/__tests__/queries.test.js
